moved stuff to the improver_tests init
